+++ userspace/archive/tar.c
@@ -1,4 +1,4 @@
-#include "lib/libstar.h"
+#include "lib/libtar.h"
 #include <stdio.h>
 #include <stdlib.h>
