+++ userspace/archive/unzip.c
@@ -64,7 +64,7 @@
 #  define CHDIR(d) chdir(d)
 #endif
 
-#include "lib/unzip.h"
+#include "lib/libunzip.h"
 
 #define WRITEBUFFERSIZE (8192)
 #define MAXFILENAME     (256)
