+++ userspace/lib/libzip.h
